requirements.txt
tensorflow
opencv-python
flask
Pillow
numpy
